Lots of changes, read description
Created and integrated the following debug features:

Inline break points: Allows you to insert breakpoints into the code

Assertions: Enables evaluation of expressions, and causes a executes a
break point of the evauluation fails. Useful for error checking.

Fatal error exit path: Allows you to print an error message to the
screen then exit the game.

Log file system: Allows you to print to the console and write to a log
file simultaneously with the log class. Has stream insertion operator
overloaded for ease of use.